requirements.txt
# requirements.txt
python-dotenv
openai
streamlit